small DX improvements
