+++ picker-starter/src/components/ItemPicker/utils/mixin/showCursorPagination.ts
@@ -1,5 +1,5 @@
 import { State } from '../state/types/State'
-import { isCursorPaginatedResult } from '@/domain'
+import { isCursorPaginatedResult } from '@/core'
 
 export const showCursorPagination = (state: State): boolean =>
   isCursorPaginatedResult(state.pageInfo) &&
