+++ picker-starter/src/components/ItemPicker/utils/mixin/showPagePagination.ts
@@ -1,5 +1,5 @@
 import { State } from '../state/types/State'
-import { isPagePaginatedResult } from '@/domain'
+import { isPagePaginatedResult } from '@/core'
 
 export const showPagePagination = (state: State): boolean =>
   isPagePaginatedResult(state.pageInfo) &&
