+++ picker-starter/src/options-schema.ts
@@ -3,6 +3,8 @@ import Joi from 'joi'
 export type MockPluginOptions = {
   // The maximum number of items that can be added
   limit?: string
+  // Limit the end-user from only selecting either categories or products
+  selectOnly?: string
 }
 
 const optionsSchemaMixin: Joi.StrictSchemaMap<MockPluginOptions> = {
@@ -12,6 +14,7 @@ const optionsSchemaMixin: Joi.StrictSchemaMap<MockPluginOptions> = {
     .messages({
       'string.pattern.base': '"limit" must be an integer greater than 0',
     }),
+  selectOnly: Joi.string().valid('category', 'product').optional(),
 }
 
 export const optionsSchema = Joi.object<MockPluginOptions, true>({
@@ -20,4 +23,5 @@ export const optionsSchema = Joi.object<MockPluginOptions, true>({
 
 export const optionsExample: MockPluginOptions = {
   limit: '10',
+  selectOnly: 'product',
 }
